# Minor perf improvements in parseTreeInfo()