```sh
$ benchstat old.txt new.txt 
goos: darwin
goarch: arm64
pkg: github.com/maxmind/xgb2code/gen
                 │   old.txt   │               new.txt               │
                 │   sec/op    │   sec/op     vs base                │
ParseTreeInfo-10   584.5n ± 1%   231.9n ± 1%  -60.32% (p=0.000 n=10)

                 │  old.txt   │              new.txt               │
                 │    B/op    │    B/op     vs base                │
ParseTreeInfo-10   872.0 ± 0%   640.0 ± 0%  -26.61% (p=0.000 n=10)

                 │   old.txt   │              new.txt               │
                 │  allocs/op  │ allocs/op   vs base                │
ParseTreeInfo-10   18.000 ± 0%   1.000 ± 0%  -94.44% (p=0.000 n=10)
```

<details>
<summary>before</summary>

```sh
$ go test -run=^$ -bench=. -benchmem -count=10 ./gen | tee old.txt 
goos: darwin
goarch: arm64
pkg: github.com/maxmind/xgb2code/gen
BenchmarkParseTreeInfo-10    	 2012709	       581.7 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2064264	       586.4 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2045088	       587.9 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2061264	       583.6 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2076304	       586.3 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2064022	       583.2 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2058519	       584.5 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 1977475	       584.5 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2059624	       588.7 ns/op	     872 B/op	      18 allocs/op
BenchmarkParseTreeInfo-10    	 2048594	       583.1 ns/op	     872 B/op	      18 allocs/op
PASS
ok  	github.com/maxmind/xgb2code/gen	18.236s
```

</details>

<details>
<summary>after</summary>

```sh
$ go test -run=^$ -bench=. -benchmem -count=10 ./gen | tee new.txt
goos: darwin
goarch: arm64
pkg: github.com/maxmind/xgb2code/gen
BenchmarkParseTreeInfo-10    	 4891856	       232.7 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5157418	       232.1 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5075401	       231.5 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5138415	       231.6 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5186865	       233.6 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5210715	       231.8 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5193562	       231.5 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 4877660	       231.3 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5125712	       232.1 ns/op	     640 B/op	       1 allocs/op
BenchmarkParseTreeInfo-10    	 5221425	       234.9 ns/op	     640 B/op	       1 allocs/op
PASS
ok  	github.com/maxmind/xgb2code/gen	14.573s
```

</details>